requirements.txt
django>=4.2
djangorestframework
django-cors-headers
Pillow
aiogram>=3.7
python-dotenv
aiohttp
